ncu-ci: refactor CI parsers
