+++ bin/ncu-ci
@@ -3,16 +3,19 @@
 'use strict';
 
 const {
-  PRBuild, BenchmarkRun, CommitBuild, parseJobFromURL,
-  constants, JobParser, listBuilds
-  // , jobCache
-} = require('../lib/ci');
-const clipboardy = require('clipboardy');
-const { writeJson } = require('../lib/file');
+  JobParser,
+  parseJobFromURL,
+  CI_TYPES_KEYS: {
+    PR, COMMIT, BENCHMARK
+  }
+} = require('../lib/ci/ci_type_parser');
 
 const {
-  PR, COMMIT, BENCHMARK
-} = constants;
+  PRBuild, BenchmarkRun, CommitBuild, listBuilds
+  // , jobCache
+} = require('../lib/ci/ci_result_parser');
+const clipboardy = require('clipboardy');
+const { writeJson } = require('../lib/file');
 
 const { runPromise } = require('../lib/run');
 const auth = require('../lib/auth');
