+++ lib/ci/ci_failure_parser.js
