+++ lib/pr_checker.js
@@ -17,8 +17,12 @@ const {
   CONFLICTING
 } = require('./mergeable_state');
 
-const { JobParser, CI_TYPES, constants } = require('./ci');
-const { COMMIT, PR, LITE_PR, LITE_COMMIT } = constants;
+const {
+  JobParser,
+  CI_TYPES,
+  isLiteCI,
+  isFullCI
+} = require('./ci/ci_type_parser');
 
 class PRChecker {
   /**
@@ -187,8 +191,8 @@ class PRChecker {
   }
 
   hasFullOrLiteCI(ciMap) {
-    return (ciMap.has(COMMIT) || ciMap.has(PR) ||
-            ciMap.has(LITE_PR) || ciMap.has(LITE_COMMIT));
+    const cis = [...ciMap.keys()];
+    return cis.find(ci => isFullCI(ci) || isLiteCI(ci));
   }
 
   // TODO: we might want to check CI status when it's less flaky...
